% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_coordinates.R
\name{add_coordinates}
\alias{add_coordinates}
\title{Add location coordinates}
\usage{
add_coordinates(ef, loc_column = "country", lon_lat_columns = c("lon",
  "lat"), overwrite = FALSE)
}
\arguments{
\item{ef}{An \code{epiflows} object.}

\item{loc_column}{Name of the column where location names are stored
(default: "country").}

\item{lon_lat_columns}{Names of the appended columns with longitudes
and latitudes, respectively (default: "lon" and "lat").}

\item{overwrite}{If TRUE, retrieves all geocodes, even those already
retrieved. If FALSE (default), overwrites only NAs.}
}
\value{
An updated \code{epiflows} object.
}
\description{
Adds longitude/latitude values to location data within an epiflows object.
Coordinates are added to object's location linelist as `lon` and `lat`
columns.
}
\examples{
flows <- make_epiflows(Mex_travel_2009[[2]], Mex_travel_2009[[1]])
flows <- add_coordinates(flows)
flows$linelist

}
\author{
Pawel Piatkowski
}
