% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.SpatialLinesDataFrame.R
\name{as.SpatialLinesDataFrame}
\alias{as.SpatialLinesDataFrame}
\alias{as.SpatialLinesDataFrame.epiflows}
\title{Convert to SpatialLinesDataFrame class}
\usage{
as.SpatialLinesDataFrame(x)

\method{as.SpatialLinesDataFrame}{epiflows}(x)
}
\arguments{
\item{x}{an epiflows object}
}
\value{
an object of class [sp::SpatialLinesDataFrame]
}
\description{
Convert to SpatialLinesDataFrame class
}
\examples{
data("Brazil_epiflows")
data("YF_coordinates")
ef <- add_coordinates(Brazil_epiflows, YF_coordinates[-1])
ef2 <- epicontacts::thin(ef[j = c("Espirito Santo", "Italy"), contacts = "both"])
as.SpatialLinesDataFrame(ef2)
}
