% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiflows-package.r
\docType{package}
\name{epiflows}
\alias{epiflows-package}
\alias{epiflows}
\title{epiflows}
\description{
\code{epiflows} is a package for predicting and visualising spread of infectious
diseases based on flows between geographical locations, e.g., countries.
\code{epiflows} provides functions for calculating spread estimates, handling flow
data, and visualization.
}
\details{
The main functions of the package are \code{\link[=make_epiflows]{make_epiflows()}} for
constructing an epiflows object and \code{\link[=estimate_risk_spread]{estimate_risk_spread()}} to estimate
potential spread. Visualization is available for maps, networks, and grids.

For details, see the vignettes: \cr
\code{vignette("introduction", package = "epiflows")}\cr
\code{vignette("epiflows-class", package = "epiflows")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.repidemicsconsortium.org/epiflows/}
  \item \url{https://github.com/reconhub/epiflows}
  \item Report bugs at \url{https://github.com/reconhub/epiflows/issues}
}

}
\author{
\strong{Maintainer}: Pawel Piatkowski \email{pawel.piatkowski@posteo.net} (\href{https://orcid.org/0000-0002-0822-5592}{ORCID})

Authors:
\itemize{
  \item Paula Moraga \email{p.e.moraga-serrano@lancaster.ac.uk} (\href{https://orcid.org/0000-0001-5266-0201}{ORCID})
  \item Thibaut Jombart \email{thibautjombart@gmail.com}
  \item VP Nagraj \email{vpnagraj@virginia.edu}
  \item Zhian N. Kamvar \email{zkamvar@gmail.com} (\href{https://orcid.org/0000-0003-1458-7108}{ORCID})
  \item Salla E. Toikkanen \email{Salla.e.toikkanen@gmail.com}
}

Other contributors:
\itemize{
  \item Isobel Blake \email{isobel.blake@imperial.ac.uk} [contributor, data contributor]
}

}
