% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_risk_spread.R
\name{estimate_risk_spread}
\alias{estimate_risk_spread}
\title{Travel-related disease cases spreaded to other locations from an infectious location}
\usage{
estimate_risk_spread(ef, ef_time = "time_window", location_code,
  time_window_days, num_infec_cases_in_time_window, avg_length_stay_days,
  distribution_incubation, params_incubation, distribution_infectious,
  params_infectious, num_simulations = 1000)
}
\arguments{
\item{ef}{an \code{epiflows} object. It contains the number of travellers
to and from other locations}

\item{ef_time}{character string denoting the time window of the population flows in the epiflows object.
Options are "time_window", "annual", "quarterly"}

\item{location_code}{a character string denoting the location code}

\item{time_window_days}{number of days between the first and last
disease case in infectious location}

\item{num_infec_cases_in_time_window}{cumulative number of cases
in infectious location in time window}

\item{avg_length_stay_days}{average length of stay in days of travellers from
other locations visiting the infectious location. This can be a common number
for all locations or a vector with different numbers for each location}

\item{distribution_incubation}{random generation distribution of the incubation period}

\item{params_incubation}{vector with the parameters of the incubation distribution}

\item{distribution_infectious}{random generation distribution of the infectious period}

\item{params_infectious}{vector with the parameters of the infectious distribution}

\item{num_simulations}{number of simulations from the incubation and infectious distributions}
}
\value{
data.frame with the mean and 95% confidence interval of the number
of cases spread to each country
}
\description{
Calculates the mean and 95% confidence interval of the estimated number of
disease cases that could potentially seed a disease outbreak in the locations
they are travelling to, comprising exportations (infected residents of the infectious
location travelling abroad during the incubation or infectious period), and
importations (international tourists infected by the disease during
their stay in the infectious location and returning to their home location).
The mean and 95% confidence intervals are obtained by numerically sampling
num_simulations times from the incubation and infectious period distributions.
}
\details{
parameters \code{distribution_incubation} and \code{distribution_infectious} denote the random generation distributions of the
incubation and infectious periods distributions, respectively.
Thes can be specified with the name of an R function or with a function defined by the user with parameters
\code{num_simulations}, \code{parameter1}, \code{parameter2}, etc.
Examples: \code{rnorm(n, mean, sd)}, \code{rlnorm(n, meanlog, sdlog)}, \code{rgamma(n, shape, rate)}, \code{rweibull(n, shape, scale)}, \code{rexp(n, rate)}
}
\examples{
ef <- do.call(make_epiflows, Mex_travel_2009)
num_countries <- nrow(ef$locationsdata)
avg_length_stay_days <- rpois(num_countries, 30)
res <- estimate_risk_spread(
  ef = ef,
  ef_time = "time_window",
  location_code = "MEX",
  time_window_days = 365,
  num_infec_cases_in_time_window = 1000,
  avg_length_stay_days = avg_length_stay_days,
  distribution_incubation = rlnorm,
  params_incubation = c(4.6, sqrt(2.7)),
  distribution_infectious = rnorm,
  params_infectious = c(4.5, sqrt(0.6)),
  num_simulations = 1000
)
head(res)

}
\author{
Paula Moraga
}
