% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_number_cases_spread.R
\name{fn_number_cases_spread}
\alias{fn_number_cases_spread}
\title{Travel-related disease cases spreaded to other countries from an infectious country}
\usage{
fn_number_cases_spread(x, code, time_window_days,
  num_infec_cases_in_time_window, avg_length_stay_days, mean_incubation,
  var_incubation, mean_infectious, var_infectious)
}
\arguments{
\item{x}{an \code{epiflows} object}

\item{code}{a character string denoting location (country) code}

\item{time_window_days}{number of days between the first and last confirmed
disease cases in infectious country}

\item{num_infec_cases_in_time_window}{cumulative number of confirmed cases
reported in infectious country in time window W}

\item{avg_length_stay_days}{average length of stay in days of travellers from
other countries visiting the infectious country.
This can be an number that is common for all countries or a vector with
different numbers for each country.}

\item{mean_incubation}{mean of the incubation lognormal distribution}

\item{var_incubation}{variance of the incubation lognormal distribution}

\item{mean_infectious}{mean of the infectious normal distribution}

\item{var_infectious}{variance of the infectious normal distribution}
}
\value{
data.frame with the mean and 95% confidence interval of the number
of cases spread to each country
}
\description{
Calculates the mean and 95% confidence interval of the estimated number of
disease cases that could potentially seed a disease outbreak in the countries
they are travelling to, comprising infected residents of the infectious
country travelling abroad during the incubation or infectious period
(exportations) and international tourists infected by the disease during
their stay in the infectious country and returning to their home country
(importations).
The mean and 95% confidence intervals are obtained by numerically sampling
10,000 times the incubation and infectious period distributions.
}
\examples{
flows <- make_epiflows(Mex_travel_2009[[2]], Mex_travel_2009[[1]])
num_countries <- nrow(flows$linelist)
avg_length_stay_days <- rpois(num_countries, 30)
fn_number_cases_spread(
  flows,
  "MEX",
  time_window_days = 365,
  num_infec_cases_in_time_window = 1000,
  avg_length_stay_days = avg_length_stay_days,
  mean_incubation = 4.6,
  var_incubation = 2.7,
  mean_infectious = 4.5,
  var_infectious = 0.6
)

}
\author{
Paula Moraga
}
