% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_flow_data.R
\name{get_flow_data}
\alias{get_flow_data}
\title{Get flow data for a given location}
\usage{
get_flow_data(ef, code, direction = "both")
}
\arguments{
\item{ef}{An \code{epiflows} object.}

\item{code}{A character string denoting location code.}

\item{direction}{If "to" or "from", the function returns a vector
of flows to or from the location, respectively.
If set to "both" - a two-element list with flows both to and from
the location.}
}
\description{
Returns a vector (if direction is "both") or a list of 2 elements
(if direction is "to" or "from") to and/or from the specified location.
}
\examples{
flows <- make_epiflows(Mex_travel_2009)
get_flow_data(flows, "MEX", direction = "both")

}
\author{
Pawel Piatkowski
}
