% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_coordinates.R
\name{get_geocoding_function}
\alias{get_geocoding_function}
\title{Get geocoding function}
\usage{
get_geocoding_function(backend)
}
\arguments{
\item{backend}{Either a function, or a character string pointing to one of
the supported geocoding backends. If `backend` is a function, it should
accept a vector of character strings as input (optional parameters can also
be passed from \code{add_coordinates()}) and return a data frame or a tibble.}
}
\value{
A function object.
}
\description{
Retrieves a function used by \link{add_coordinates} to geocode coordinates
for locations from an `epiflows` object.
}
