% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_locations.R
\name{get_locations}
\alias{get_locations}
\alias{get_locations.epiflows}
\title{Access flow data}
\usage{
get_locations(x, ...)

\method{get_locations}{epiflows}(x)
}
\arguments{
\item{x}{An \code{epiflows} object.}
}
\value{
A \code{data.frame} with at least 1 column called \code{id},
  specifying the id of the location used in the `flows` data frame.
}
\description{
This accessor extract location data from an \code{epiflows} object.
\code{get_locations} is a generic with a method defined for \code{epiflows}
objects.
}
\author{
Thibaut Jombart, Zhian Kamvar
}
