% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vars.R
\name{get_vars}
\alias{get_vars}
\alias{get_vars.epiflows}
\alias{get_coords}
\alias{get_coords.epiflows}
\title{Access flow data}
\usage{
get_vars(x, ...)

\method{get_vars}{epiflows}(x, what = NULL)

get_coords(x)

\method{get_coords}{epiflows}(x)
}
\arguments{
\item{x}{An \code{epiflows} object.}

\item{what}{a valid character string specifying the variable desired. If
\code{NULL} (default), the names of the available vars will be returned.}
}
\value{
A data frame with the variables requested
}
\description{
This accessor extracts variables from the \code{locations} data frame in an
\code{epiflow} object. \code{get_vars} is a generic with a method defined for
\code{epiflows} objects.
}
\details{
The
}
\author{
Thibaut Jombart, Zhian Kamvar
}
