% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_epiflows.R
\name{grid_epiflows}
\alias{grid_epiflows}
\title{Visualise epidemic flows using a grid}
\usage{
grid_epiflows(x, ...)
}
\arguments{
\item{x}{An \code{epiflows} object.}
}
\description{
This grid plot shows flows between locations by positioning origins and
destination in y and x axes, respectively. This function is called when
plotting \code{epiflows} with \code{type = "grid"}.
}
\author{
Thibaut Jombart
}
