% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Mex_travel_2009}
\alias{Mex_travel_2009}
\title{Travel data to and from Mexico in 2009}
\format{A list of 2 elements
\describe{
  \item{1}{An unsymmetrical matrix (data frame object) of ...}
  \item{2}{Country metadata (data frame object) ...}
}}
\usage{
Mex_travel_2009
}
\description{
<Description>
}
\keyword{datasets}
