% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_coordinates.R
\name{nominatim_backend_function}
\alias{nominatim_backend_function}
\title{Nominatim geocoding backend}
\usage{
nominatim_backend_function(query, ...)
}
\arguments{
\item{query}{Geocoding query.}

\item{...}{Optional parameters.}
}
\value{
A data frame with two columns: `lat` and `lon`.
}
\description{
Wrapper for Nominatim geocoding function provided by `nominatim` package.
See the documentation at https://github.com/hrbrmstr/nominatim for details.
}
