% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_flows.R
\name{random_flows}
\alias{random_flows}
\title{Generate random flows inputs}
\usage{
random_flows(n, pool = letters)
}
\value{
A \code{data.frame} with 3 columns, which can serve as input to the
  constructor:

\itemize{

 \item \code{from}: origin of the flow

 \item \code{to}: destination of the flow

 \item \code{n}: magnitude of the flow - can be a number of passengers per
 unit of time, a rate, a probability of migration

}
}
\description{
This internal function generates random flows on a scale of 1 to 100 from a
pool of locations.
}
\author{
Thibaut Jombart
}
