% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handling.R
\name{[.epiflows}
\alias{[.epiflows}
\title{Subset `epiflows` objects}
\usage{
\method{[}{epiflows}(ef, i, ...)
}
\arguments{
\item{ef}{An \code{epiflows} object.}

\item{i}{A character vector of one or more location codes.}

\item{...}{Additional parameters (not used).}
}
\value{
An \code{epiflows} object.
}
\description{
Returns a subset of an epiflows object based on location code(s).
}
\examples{
flows <- do.call(make_epiflows, Mex_travel_2009)
flows["MEX"]

}
\author{
Pawel Piatkowski
}
