% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_epiflows.R
\name{vis_epiflows}
\alias{vis_epiflows}
\title{Visualise epidemic flows using visNetwork}
\usage{
vis_epiflows(x, ...)
}
\arguments{
\item{x}{An \code{epiflows} object.}

\item{...}{Further arguments to be passed on to
\code{\link[epicontacts:vis_epicontacts]{epicontacts::vis_epicontacts()}}}
}
\description{
This function shows flows between locations using a dynamic network
visualisation implemented in the package \pkg{visNetwork}, calling the
function \code{\link[epicontacts:vis_epicontacts]{epicontacts::vis_epicontacts()}} of the \pkg{epicontacts}
package.
}
\author{
Thibaut Jombart
}
