% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiflows-class.R
\name{epiflows}
\alias{epiflows}
\title{Read linelist and contact data}
\usage{
epiflows(...)
}
\arguments{
\item{...}{named character or integers}

\item{linelist}{a \link{data.frame} with at least one column providing unique
patient identifiers}

\item{contacts}{a \link{data.frame} that needs at least two columns
indicating patients between which cases take place; these need not be
referenced in the linelist}

\item{id}{an index or name indicating which column in \code{linelist}
contains unique identifiers; default is first column in \code{linelist}
data frame}

\item{from}{an index or name indicating which column in \code{contacts}
contains the first region}

\item{to}{an index or name indicating which column in \code{contacts}
contains the second region}

\item{n}{an index or name indicating which column in \code{contacts}
contains the number of cases flowing between the "from" region and the
"to" region}
}
\value{
An \code{epicontacts} object in list format with three elements:

\itemize{
\item \code{linelist}: data.frame of cases with first column 'id'
containing character vector of unique identifiers

\item \code{contacts}: data.frame of contacts with first two columns named
'from' and 'to' indicating unique pairs of contact between individuals

\item \code{directed}: indicator as to whether or not the contacts are to be
considered directed or not
}
}
\description{
This function reads data stored as data.frame containing linelist (case
information, where each row corresponds to a unique patient), and contacts
between patients. Common identifiers should be used in the two data sources
for matching to be achieved.
}
\details{
An \code{epicontacts} object can be created from two components:
\itemize{
\item a linelist provided as a \code{data.frame} where columns are
different variables describing cases, and where each row is a different case.
and a contact list.

\item a contact list provided as a \code{data.frame} where each row contains
unique pairs of contacts with unique features of contact in columns. The line
list and contact list should share an identification scheme for individuals.
}
}
\examples{
data(YF_Brazil)
from     <- as.data.frame.table(YF_Brazil$T_D)
to       <- as.data.frame.table(YF_Brazil$T_O)[c(2,1,3)]
contacts <- rbind(setNames(from, c("from", "to", "n")), 
                  setNames(to, c("from", "to", "n")))
linelist <- YF_Brazil$states
others   <- setdiff(contacts$to, YF_Brazil$states$location_code)
linelist <- merge(linelist, 
                  data.frame(location_code = others),
                  by = "location_code", all = TRUE)
ef <- epiflows(linelist, contacts, pop_size = "num_cases_time_window")
ef
# Access variable information
ef$linelist[ef$vars$pop_size]
}
\references{
\url{http://foodborne.unl.edu/public/role/epidemiologist/lineLists.html}
}
\author{
Thibaut Jombart (\email{thibautjombart@gmail.com})
}
