% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiflows-package.r
\docType{package}
\name{epiflows}
\alias{epiflows}
\title{epiflows}
\description{
\code{epiflows} is a package for predicting and visualising spread of infectious
diseases based on flows between geographical locations, e.g., countries.
\code{epiflows} provides functions for calculating spread estimates, handling flow
data, and visualization.
}
\details{
The main functions of the package are \code{\link[=make_epiflows]{make_epiflows()}} for
constructing an epiflows object and \code{\link[=estimate_risk_spread]{estimate_risk_spread()}} to estimate
potential spread. Visualization is available for maps, networks, and grids.

For details, see the vignettes: \cr
\code{vignette("introduction", package = "epiflows")}\cr
\code{vignette("epiflows-class", package = "epiflows")}
}
