% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_flows.R
\name{get_flows}
\alias{get_flows}
\alias{get_flows.epiflows}
\title{Access flow data}
\usage{
get_flows(x, ...)

\method{get_flows}{epiflows}(x)
}
\arguments{
\item{x}{An \code{epiflows} object.}
}
\value{
A \code{data.frame} with 3 columns:

\itemize{

 \item \code{from}: origin of the flow

 \item \code{to}: destination of the flow

 \item \code{flow}: magnitude of the flow - can be a number of passengers per
 unit of time, a rate, a probability of migration

}
}
\description{
This accessor extract flow data from an \code{epiflows}
object. \code{get_flows} is a generic with a method defined for
\code{epiflows} objects.
}
\author{
Thibaut Jombart
}
