% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_id.R
\name{get_id}
\alias{get_id}
\title{Access population identifiers in epiflows objects}
\usage{
get_id(x, ...)
}
\arguments{
\item{x}{an epiflows object}

\item{...}{arguments passed on to \code{\link[epicontacts:get_id]{epicontacts::get_id()}}}
}
\value{
a character vector of population IDs
}
\description{
this will return the unique population ID for your epiflows object.
}
\examples{
data("Brazil_epiflows")
get_id(Brazil_epiflows)

}
\seealso{
\code{\link[epicontacts:get_id]{epicontacts::get_id()}}; \code{\link[=get_vars]{get_vars()}}; \code{\link[=get_pop_size]{get_pop_size()}}; \code{\link[=global_vars]{global_vars()}}
}
\author{
Zhian N. Kamvar
}
