% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pop_size.R
\name{get_pop_size}
\alias{get_pop_size}
\alias{get_pop_size.epiflows}
\title{Get population size for each entry in locations}
\usage{
get_pop_size(x)

\method{get_pop_size}{epiflows}(x)
}
\arguments{
\item{x}{an epiflows object}
}
\value{
a named vector of population sizes
}
\description{
Get population size for each entry in locations
}
\examples{
data("Brazil_epiflows")
get_pop_size(Brazil_epiflows)
}
\seealso{
\code{\link[=get_vars]{get_vars()}}, \code{\link[=global_vars]{global_vars()}}
}
