% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vars.R
\name{get_vars}
\alias{get_vars}
\alias{get_vars.epiflows}
\alias{get_coords}
\alias{get_coords.epiflows}
\title{Access flow data}
\usage{
get_vars(x, ...)

\method{get_vars}{epiflows}(x, what = NULL, id = TRUE)

get_coords(x, ...)

\method{get_coords}{epiflows}(x, location = NULL)
}
\arguments{
\item{x}{An `epiflows` object.}

\item{what}{a valid character string specifying the variable desired. If
`NULL` (default), the names of the available vars will be returned.}

\item{id}{a logical. If `TRUE` (default), the `id` column of the locations
will be the first column of the data frame. if `FALSE`, the variable will
be returned without identifiers.}

\item{location}{a character specifying a single location to return as a vector of
coordinates. You cannot specify multiple locations with this parameter.
Defaults to `NULL`, indicating all locations.}
}
\value{
A data frame with the variables requested
}
\description{
This accessor extracts variables from the \code{locations} data frame in an
\code{epiflow} object. \code{get_vars} is a generic with a method defined for
\code{epiflows} objects.
}
\details{
The function \code{get_coords()} is equivalent to \code{get_vars(x, "coords", id = TRUE)}
}
\author{
Thibaut Jombart, Zhian Kamvar
}
