% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vars.R
\name{get_vars}
\alias{get_vars}
\alias{get_vars.epiflows}
\title{Access flow data}
\usage{
get_vars(x, ...)

\method{get_vars}{epiflows}(x, y = NULL)
}
\arguments{
\item{x}{An \code{epiflows} object.}

\item{y}{a valid character string specifying the variable desired. If
\code{NULL} (default), the names of the available vars will be returned.}
}
\value{
A data frame with one
}
\description{
This accessor extract location data from an \code{epiflows}
object. \code{get_vars} is a generic with a method defined for
\code{epiflows} objects.
}
\author{
Thibaut Jombart, Zhian Kamvar
}
