% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coordinates.R, R/get_pop_size.R,
%   R/get_vars.R
\name{get_coordinates}
\alias{get_coordinates}
\alias{get_coordinates.epiflows}
\alias{get_pop_size}
\alias{get_pop_size.epiflows}
\alias{get_vars}
\alias{get_vars.epiflows}
\title{Access location metadata}
\usage{
get_coordinates(x, ...)

\method{get_coordinates}{epiflows}(x, location = NULL, ...)

get_pop_size(x)

\method{get_pop_size}{epiflows}(x)

get_vars(x, ...)

\method{get_vars}{epiflows}(x, what = NULL, id = TRUE, vector = FALSE,
  ...)
}
\arguments{
\item{x}{An \code{epiflows} object.}

\item{...}{not currently used.}

\item{location}{a character specifying a single location to return as a vector of
coordinates. You cannot specify multiple locations with this parameter.
Defaults to `NULL`, indicating all locations.}

\item{what}{a valid character string specifying the variable desired. If
`NULL` (default), the names of the available vars will be returned.}

\item{id}{a logical. If `TRUE` (default), the `id` column of the locations
will be the first column of the data frame. if `FALSE`, the variable will
be returned with identifiers as row names.}

\item{vector}{if `TRUE` the result will be coerced into a vector (or a matrix in the case of coordinates)}
}
\value{
a named vector of population sizes

A data frame with the variables requested
}
\description{
This accessor extracts variables from the \code{locations} data frame in an
\code{epiflow} object. \code{get_vars} is a generic with a method defined for
\code{epiflows} objects.
}
\details{
The function \code{get_coordinates()} is equivalent to \code{get_vars(x, "coordinates", id = TRUE)}
}
\examples{

data("Brazil_epiflows")
get_vars(Brazil_epiflows, "duration_stay")
get_vars(Brazil_epiflows, "duration_stay", vector = TRUE)
get_pop_size(Brazil_epiflows)
}
\author{
Thibaut Jombart, Zhian Kamvar
}
