% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_epiflows.R
\name{grid_epiflows}
\alias{grid_epiflows}
\title{Visualise epidemic flows using a grid}
\usage{
grid_epiflows(x, color_by = c("from", "to", "none"), ...)
}
\arguments{
\item{x}{An \code{epiflows} object.}

\item{color_by}{A character string indicating if flows should be colored by
origin (\code{from}) or destination (\code{to}).}

\item{...}{arguments passed on to \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}}
}
\description{
This grid plot shows flows between locations by positioning origins and
destination in y and x axes, respectively. This function is called when
plotting \code{epiflows} with \code{type = "grid"}.
}
\author{
Thibaut Jombart
}
