% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_epiflows.R
\name{make_epiflows}
\alias{make_epiflows}
\title{Make an `epiflows` object}
\usage{
make_epiflows(to, from, code, locationsdata)
}
\arguments{
\item{to}{A named vector of flows to the specified location.
A one-row data frame or a list is also allowed.}

\item{from}{A named vector of flows from the specified location
(see above).}

\item{code}{Character string; location code of the origin.}

\item{locationsdata}{A data frame of location data. Locations should be
uniquely identified by codes from the `code` column.}
}
\value{
An \code{epiflows} object.
}
\description{
Takes a list of location data and a matrix of flows, and creates
an object of class `epiflows`. An error is thrown if the input is invalid.
}
\examples{
# TBD

}
\author{
Pawel Piatkowski
}
