% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_epiflows.R
\name{make_epiflows}
\alias{make_epiflows}
\title{Make an `epiflows` object}
\usage{
make_epiflows(flows, locationsdata)
}
\arguments{
\item{flows}{A numeric matrix of flows between locations. Row and column
names must denote location codes.}

\item{locationsdata}{A data frame of location data. Locations should be uniquely
identified by codes from the `code` column.}
}
\value{
An \code{epiflows} object.
}
\description{
Takes a list of location data and a matrix of flows, and creates
an object of class `epiflows`. An error is thrown if the input is invalid.
}
\examples{
input_data <- Mex_travel_2009
make_epiflows(input_data[[1]], input_data[[2]])

}
\author{
Pawel Piatkowski
}
