% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.epiflows.R
\name{plot.epiflows}
\alias{plot.epiflows}
\title{Plot a flow map}
\usage{
\method{plot}{epiflows}(ef, origin, title = sprintf("Flows from \%s",
  origin_name), loc_column = "country", lon_lat_columns = c("lon", "lat"))
}
\arguments{
\item{ef}{An \code{epiflows} object.}

\item{origin}{Code of the location of origin.}

\item{title}{Plot title.}

\item{loc_column}{Name of the column where location names are stored
(default: "country").}

\item{lon_lat_columns}{Names of the columns with longitudes and latitudes,
respectively (default: "lon" and "lat").}
}
\value{
A \code{leaflet} object
}
\description{
Returns an interactive map of population flows from the specified
location of origin.
}
\author{
Paula Moraga, Pawel Piatkowski

flows <- do.call(make_epiflows, Mex_travel_2009)
flows <- add_coordinates(flows)
plot(flows, "MEX")
}
