% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.epiflows.R
\name{print.epiflows}
\alias{print.epiflows}
\title{Print method from epiflows objects}
\usage{
\method{print}{epiflows}(x, ...)
}
\arguments{
\item{x}{An \code{epiflows} object.}

\item{...}{Additional parameters (not used).}
}
\description{
Displays a short summary of an \code{epiflows} object.
}
\examples{
data("Brazil_epiflows")
print(Brazil_epiflows)

}
\author{
Zhian N. Kamvar, Thibaut Jombart
}
