% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_flows.R
\name{random_flows}
\alias{random_flows}
\alias{random_flows.epiflows}
\title{Access flow data}
\usage{
random_flows(x, ...)

\method{random_flows}{epiflows}(x, n = nrow(x$contacts), ...)
}
\arguments{
\item{x}{An \code{epiflows} object.}
}
\value{
A \code{data.frame} with 3 columns:

\itemize{

 \item \code{from}: origin of the flow

 \item \code{to}: destination of the flow

 \item \code{flow}: magnitude of the flow - can be a number of passengers per
 unit of time, a rate, a probability of migration

}
}
\description{
This accessor extract random flow data from an \code{epiflows}
object. \code{random_flows} is a generic with a method defined for
\code{epiflows} objects.
}
\author{
Thibaut Jombart
}
